buffalo/requirements.txt
pandas>=1.5.2
polygon-api-client>=1.7.1
openpyxl>=3.0.10